@@ -1,30 +1,31 @@
+//
+// SNMP Protobuf.
+//
+// A protobuf interface to SNMP functions.
+
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.36.4
+// 	protoc        v5.29.3
 // source: snmp.proto
 
-package thebinary_snmp
+package protobuf
 
 import (
-	context "context"
-	fmt "fmt"
-	proto "github.com/golang/protobuf/proto"
-	grpc "google.golang.org/grpc"
-	codes "google.golang.org/grpc/codes"
-	status "google.golang.org/grpc/status"
-	math "math"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
+	unsafe "unsafe"
 )
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
-
-//
 // Asn1Ber Enum Type definitions
 type Asn1BER int32
 
@@ -53,204 +54,290 @@ const (
 	Asn1BER_EndOfMibView      Asn1BER = 130
 )
 
-var Asn1BER_name = map[int32]string{
-	0: "EndOfContents",
-	// Duplicate value: 0: "UnknownType",
-	1:   "Boolean",
-	2:   "Integer",
-	3:   "BitString",
-	4:   "OctetString",
-	5:   "Null",
-	6:   "ObjectIdentifier",
-	7:   "ObjectDescription",
-	64:  "IPAddress",
-	65:  "Counter32",
-	66:  "Gauge32",
-	67:  "TimeTicks",
-	68:  "Opaque",
-	69:  "NsapAddress",
-	70:  "Counter64",
-	71:  "Uinteger32",
-	120: "OpaqueFloat",
-	121: "OpaqueDouble",
-	128: "NoSuchObject",
-	129: "NoSuchInstance",
-	130: "EndOfMibView",
-}
+// Enum value maps for Asn1BER.
+var (
+	Asn1BER_name = map[int32]string{
+		0: "EndOfContents",
+		// Duplicate value: 0: "UnknownType",
+		1:   "Boolean",
+		2:   "Integer",
+		3:   "BitString",
+		4:   "OctetString",
+		5:   "Null",
+		6:   "ObjectIdentifier",
+		7:   "ObjectDescription",
+		64:  "IPAddress",
+		65:  "Counter32",
+		66:  "Gauge32",
+		67:  "TimeTicks",
+		68:  "Opaque",
+		69:  "NsapAddress",
+		70:  "Counter64",
+		71:  "Uinteger32",
+		120: "OpaqueFloat",
+		121: "OpaqueDouble",
+		128: "NoSuchObject",
+		129: "NoSuchInstance",
+		130: "EndOfMibView",
+	}
+	Asn1BER_value = map[string]int32{
+		"EndOfContents":     0,
+		"UnknownType":       0,
+		"Boolean":           1,
+		"Integer":           2,
+		"BitString":         3,
+		"OctetString":       4,
+		"Null":              5,
+		"ObjectIdentifier":  6,
+		"ObjectDescription": 7,
+		"IPAddress":         64,
+		"Counter32":         65,
+		"Gauge32":           66,
+		"TimeTicks":         67,
+		"Opaque":            68,
+		"NsapAddress":       69,
+		"Counter64":         70,
+		"Uinteger32":        71,
+		"OpaqueFloat":       120,
+		"OpaqueDouble":      121,
+		"NoSuchObject":      128,
+		"NoSuchInstance":    129,
+		"EndOfMibView":      130,
+	}
+)
 
-var Asn1BER_value = map[string]int32{
-	"EndOfContents":     0,
-	"UnknownType":       0,
-	"Boolean":           1,
-	"Integer":           2,
-	"BitString":         3,
-	"OctetString":       4,
-	"Null":              5,
-	"ObjectIdentifier":  6,
-	"ObjectDescription": 7,
-	"IPAddress":         64,
-	"Counter32":         65,
-	"Gauge32":           66,
-	"TimeTicks":         67,
-	"Opaque":            68,
-	"NsapAddress":       69,
-	"Counter64":         70,
-	"Uinteger32":        71,
-	"OpaqueFloat":       120,
-	"OpaqueDouble":      121,
-	"NoSuchObject":      128,
-	"NoSuchInstance":    129,
-	"EndOfMibView":      130,
+func (x Asn1BER) Enum() *Asn1BER {
+	p := new(Asn1BER)
+	*p = x
+	return p
 }
 
 func (x Asn1BER) String() string {
-	return proto.EnumName(Asn1BER_name, int32(x))
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
 }
 
+func (Asn1BER) Descriptor() protoreflect.EnumDescriptor {
+	return file_snmp_proto_enumTypes[0].Descriptor()
+}
+
+func (Asn1BER) Type() protoreflect.EnumType {
+	return &file_snmp_proto_enumTypes[0]
+}
+
+func (x Asn1BER) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use Asn1BER.Descriptor instead.
 func (Asn1BER) EnumDescriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{0}
+	return file_snmp_proto_rawDescGZIP(), []int{0}
 }
 
-//
 // Represents SNMP OID.
 type Oid struct {
-	Oid                  string   `protobuf:"bytes,1,opt,name=oid,proto3" json:"oid,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Oid           string                 `protobuf:"bytes,1,opt,name=oid,proto3" json:"oid,omitempty"` // SNMP OID
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *Oid) Reset()         { *m = Oid{} }
-func (m *Oid) String() string { return proto.CompactTextString(m) }
-func (*Oid) ProtoMessage()    {}
+func (x *Oid) Reset() {
+	*x = Oid{}
+	mi := &file_snmp_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *Oid) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Oid) ProtoMessage() {}
+
+func (x *Oid) ProtoReflect() protoreflect.Message {
+	mi := &file_snmp_proto_msgTypes[0]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Oid.ProtoReflect.Descriptor instead.
 func (*Oid) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{0}
+	return file_snmp_proto_rawDescGZIP(), []int{0}
 }
 
-func (m *Oid) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_Oid.Unmarshal(m, b)
-}
-func (m *Oid) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_Oid.Marshal(b, m, deterministic)
-}
-func (m *Oid) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Oid.Merge(m, src)
-}
-func (m *Oid) XXX_Size() int {
-	return xxx_messageInfo_Oid.Size(m)
-}
-func (m *Oid) XXX_DiscardUnknown() {
-	xxx_messageInfo_Oid.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Oid proto.InternalMessageInfo
-
-func (m *Oid) GetOid() string {
-	if m != nil {
-		return m.Oid
+func (x *Oid) GetOid() string {
+	if x != nil {
+		return x.Oid
 	}
 	return ""
 }
 
-//
 // Represents list of SNMP Oids
 type OidList struct {
-	Oids                 []string `protobuf:"bytes,1,rep,name=oids,proto3" json:"oids,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Oids          []string               `protobuf:"bytes,1,rep,name=oids,proto3" json:"oids,omitempty"` // List of SNMP OIDs
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *OidList) Reset()         { *m = OidList{} }
-func (m *OidList) String() string { return proto.CompactTextString(m) }
-func (*OidList) ProtoMessage()    {}
+func (x *OidList) Reset() {
+	*x = OidList{}
+	mi := &file_snmp_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *OidList) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*OidList) ProtoMessage() {}
+
+func (x *OidList) ProtoReflect() protoreflect.Message {
+	mi := &file_snmp_proto_msgTypes[1]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use OidList.ProtoReflect.Descriptor instead.
 func (*OidList) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{1}
+	return file_snmp_proto_rawDescGZIP(), []int{1}
 }
 
-func (m *OidList) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_OidList.Unmarshal(m, b)
-}
-func (m *OidList) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_OidList.Marshal(b, m, deterministic)
-}
-func (m *OidList) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_OidList.Merge(m, src)
-}
-func (m *OidList) XXX_Size() int {
-	return xxx_messageInfo_OidList.Size(m)
-}
-func (m *OidList) XXX_DiscardUnknown() {
-	xxx_messageInfo_OidList.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_OidList proto.InternalMessageInfo
-
-func (m *OidList) GetOids() []string {
-	if m != nil {
-		return m.Oids
+func (x *OidList) GetOids() []string {
+	if x != nil {
+		return x.Oids
 	}
 	return nil
 }
 
-//
 // Represents a single SNMP PDU
 // consisting of oid, type and a value.
 //
 // The value is in any one of the following fields,
 // determined by the type of value it stores.
 type SnmpPDU struct {
-	Name string  `protobuf:"bytes,1,opt,name=Name,proto3" json:"Name,omitempty"`
-	Type Asn1BER `protobuf:"varint,2,opt,name=Type,proto3,enum=thebinary.snmp.Asn1BER" json:"Type,omitempty"`
+	state protoimpl.MessageState `protogen:"open.v1"`
+	Name  string                 `protobuf:"bytes,1,opt,name=Name,proto3" json:"Name,omitempty"`                              // OID
+	Type  Asn1BER                `protobuf:"varint,2,opt,name=Type,proto3,enum=thebinary.snmp.Asn1BER" json:"Type,omitempty"` // PDU Type (Asn1BER encoding type)
 	// Types that are valid to be assigned to Value:
+	//
 	//	*SnmpPDU_I32
 	//	*SnmpPDU_I64
 	//	*SnmpPDU_UI32
 	//	*SnmpPDU_UI64
 	//	*SnmpPDU_Str
-	Value                isSnmpPDU_Value `protobuf_oneof:"Value"`
-	XXX_NoUnkeyedLiteral struct{}        `json:"-"`
-	XXX_unrecognized     []byte          `json:"-"`
-	XXX_sizecache        int32           `json:"-"`
+	Value         isSnmpPDU_Value `protobuf_oneof:"Value"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *SnmpPDU) Reset()         { *m = SnmpPDU{} }
-func (m *SnmpPDU) String() string { return proto.CompactTextString(m) }
-func (*SnmpPDU) ProtoMessage()    {}
+func (x *SnmpPDU) Reset() {
+	*x = SnmpPDU{}
+	mi := &file_snmp_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *SnmpPDU) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*SnmpPDU) ProtoMessage() {}
+
+func (x *SnmpPDU) ProtoReflect() protoreflect.Message {
+	mi := &file_snmp_proto_msgTypes[2]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use SnmpPDU.ProtoReflect.Descriptor instead.
 func (*SnmpPDU) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{2}
+	return file_snmp_proto_rawDescGZIP(), []int{2}
 }
 
-func (m *SnmpPDU) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_SnmpPDU.Unmarshal(m, b)
-}
-func (m *SnmpPDU) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_SnmpPDU.Marshal(b, m, deterministic)
-}
-func (m *SnmpPDU) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_SnmpPDU.Merge(m, src)
-}
-func (m *SnmpPDU) XXX_Size() int {
-	return xxx_messageInfo_SnmpPDU.Size(m)
-}
-func (m *SnmpPDU) XXX_DiscardUnknown() {
-	xxx_messageInfo_SnmpPDU.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_SnmpPDU proto.InternalMessageInfo
-
-func (m *SnmpPDU) GetName() string {
-	if m != nil {
-		return m.Name
+func (x *SnmpPDU) GetName() string {
+	if x != nil {
+		return x.Name
 	}
 	return ""
 }
 
-func (m *SnmpPDU) GetType() Asn1BER {
-	if m != nil {
-		return m.Type
+func (x *SnmpPDU) GetType() Asn1BER {
+	if x != nil {
+		return x.Type
 	}
 	return Asn1BER_EndOfContents
+}
+
+func (x *SnmpPDU) GetValue() isSnmpPDU_Value {
+	if x != nil {
+		return x.Value
+	}
+	return nil
+}
+
+func (x *SnmpPDU) GetI32() int32 {
+	if x != nil {
+		if x, ok := x.Value.(*SnmpPDU_I32); ok {
+			return x.I32
+		}
+	}
+	return 0
+}
+
+func (x *SnmpPDU) GetI64() int64 {
+	if x != nil {
+		if x, ok := x.Value.(*SnmpPDU_I64); ok {
+			return x.I64
+		}
+	}
+	return 0
+}
+
+func (x *SnmpPDU) GetUI32() uint32 {
+	if x != nil {
+		if x, ok := x.Value.(*SnmpPDU_UI32); ok {
+			return x.UI32
+		}
+	}
+	return 0
+}
+
+func (x *SnmpPDU) GetUI64() uint64 {
+	if x != nil {
+		if x, ok := x.Value.(*SnmpPDU_UI64); ok {
+			return x.UI64
+		}
+	}
+	return 0
+}
+
+func (x *SnmpPDU) GetStr() string {
+	if x != nil {
+		if x, ok := x.Value.(*SnmpPDU_Str); ok {
+			return x.Str
+		}
+	}
+	return ""
 }
 
 type isSnmpPDU_Value interface {
@@ -258,23 +345,23 @@ type isSnmpPDU_Value interface {
 }
 
 type SnmpPDU_I32 struct {
-	I32 int32 `protobuf:"varint,3,opt,name=I32,proto3,oneof"`
+	I32 int32 `protobuf:"varint,3,opt,name=I32,proto3,oneof"` // Stores 32-bit integer
 }
 
 type SnmpPDU_I64 struct {
-	I64 int64 `protobuf:"varint,4,opt,name=I64,proto3,oneof"`
+	I64 int64 `protobuf:"varint,4,opt,name=I64,proto3,oneof"` // Stores 64-bit signed integer
 }
 
 type SnmpPDU_UI32 struct {
-	UI32 uint32 `protobuf:"varint,5,opt,name=UI32,proto3,oneof"`
+	UI32 uint32 `protobuf:"varint,5,opt,name=UI32,proto3,oneof"` // Stores 32-bit unsigned integer
 }
 
 type SnmpPDU_UI64 struct {
-	UI64 uint64 `protobuf:"varint,6,opt,name=UI64,proto3,oneof"`
+	UI64 uint64 `protobuf:"varint,6,opt,name=UI64,proto3,oneof"` // Stores 64-bit unsigned integer
 }
 
 type SnmpPDU_Str struct {
-	Str string `protobuf:"bytes,7,opt,name=Str,proto3,oneof"`
+	Str string `protobuf:"bytes,7,opt,name=Str,proto3,oneof"` // Stores string
 }
 
 func (*SnmpPDU_I32) isSnmpPDU_Value() {}
@@ -287,421 +374,244 @@ func (*SnmpPDU_UI64) isSnmpPDU_Value() {}
 
 func (*SnmpPDU_Str) isSnmpPDU_Value() {}
 
-func (m *SnmpPDU) GetValue() isSnmpPDU_Value {
-	if m != nil {
-		return m.Value
+// Represents multiple SNMP PDU
+type SnmpPDUs struct {
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Pdus          []*SnmpPDU             `protobuf:"bytes,1,rep,name=pdus,proto3" json:"pdus,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
+}
+
+func (x *SnmpPDUs) Reset() {
+	*x = SnmpPDUs{}
+	mi := &file_snmp_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *SnmpPDUs) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*SnmpPDUs) ProtoMessage() {}
+
+func (x *SnmpPDUs) ProtoReflect() protoreflect.Message {
+	mi := &file_snmp_proto_msgTypes[3]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use SnmpPDUs.ProtoReflect.Descriptor instead.
+func (*SnmpPDUs) Descriptor() ([]byte, []int) {
+	return file_snmp_proto_rawDescGZIP(), []int{3}
+}
+
+func (x *SnmpPDUs) GetPdus() []*SnmpPDU {
+	if x != nil {
+		return x.Pdus
 	}
 	return nil
 }
 
-func (m *SnmpPDU) GetI32() int32 {
-	if x, ok := m.GetValue().(*SnmpPDU_I32); ok {
-		return x.I32
+type SnmpPacket struct {
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Error         uint32                 `protobuf:"varint,1,opt,name=Error,proto3" json:"Error,omitempty"`
+	Variable      []*SnmpPDU             `protobuf:"bytes,2,rep,name=Variable,proto3" json:"Variable,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
+}
+
+func (x *SnmpPacket) Reset() {
+	*x = SnmpPacket{}
+	mi := &file_snmp_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *SnmpPacket) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*SnmpPacket) ProtoMessage() {}
+
+func (x *SnmpPacket) ProtoReflect() protoreflect.Message {
+	mi := &file_snmp_proto_msgTypes[4]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use SnmpPacket.ProtoReflect.Descriptor instead.
+func (*SnmpPacket) Descriptor() ([]byte, []int) {
+	return file_snmp_proto_rawDescGZIP(), []int{4}
+}
+
+func (x *SnmpPacket) GetError() uint32 {
+	if x != nil {
+		return x.Error
 	}
 	return 0
 }
 
-func (m *SnmpPDU) GetI64() int64 {
-	if x, ok := m.GetValue().(*SnmpPDU_I64); ok {
-		return x.I64
+func (x *SnmpPacket) GetVariable() []*SnmpPDU {
+	if x != nil {
+		return x.Variable
 	}
-	return 0
+	return nil
 }
 
-func (m *SnmpPDU) GetUI32() uint32 {
-	if x, ok := m.GetValue().(*SnmpPDU_UI32); ok {
-		return x.UI32
-	}
-	return 0
+var File_snmp_proto protoreflect.FileDescriptor
+
+var file_snmp_proto_rawDesc = string([]byte{
+	0x0a, 0x0a, 0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0e, 0x74, 0x68,
+	0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x22, 0x17, 0x0a, 0x03,
+	0x4f, 0x69, 0x64, 0x12, 0x10, 0x0a, 0x03, 0x6f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x03, 0x6f, 0x69, 0x64, 0x22, 0x1d, 0x0a, 0x07, 0x4f, 0x69, 0x64, 0x4c, 0x69, 0x73, 0x74,
+	0x12, 0x12, 0x0a, 0x04, 0x6f, 0x69, 0x64, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x04,
+	0x6f, 0x69, 0x64, 0x73, 0x22, 0xbb, 0x01, 0x0a, 0x07, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55,
+	0x12, 0x12, 0x0a, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04,
+	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x2b, 0x0a, 0x04, 0x54, 0x79, 0x70, 0x65, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x0e, 0x32, 0x17, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73,
+	0x6e, 0x6d, 0x70, 0x2e, 0x41, 0x73, 0x6e, 0x31, 0x42, 0x45, 0x52, 0x52, 0x04, 0x54, 0x79, 0x70,
+	0x65, 0x12, 0x12, 0x0a, 0x03, 0x49, 0x33, 0x32, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x48, 0x00,
+	0x52, 0x03, 0x49, 0x33, 0x32, 0x12, 0x12, 0x0a, 0x03, 0x49, 0x36, 0x34, 0x18, 0x04, 0x20, 0x01,
+	0x28, 0x03, 0x48, 0x00, 0x52, 0x03, 0x49, 0x36, 0x34, 0x12, 0x14, 0x0a, 0x04, 0x55, 0x49, 0x33,
+	0x32, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0d, 0x48, 0x00, 0x52, 0x04, 0x55, 0x49, 0x33, 0x32, 0x12,
+	0x14, 0x0a, 0x04, 0x55, 0x49, 0x36, 0x34, 0x18, 0x06, 0x20, 0x01, 0x28, 0x04, 0x48, 0x00, 0x52,
+	0x04, 0x55, 0x49, 0x36, 0x34, 0x12, 0x12, 0x0a, 0x03, 0x53, 0x74, 0x72, 0x18, 0x07, 0x20, 0x01,
+	0x28, 0x09, 0x48, 0x00, 0x52, 0x03, 0x53, 0x74, 0x72, 0x42, 0x07, 0x0a, 0x05, 0x56, 0x61, 0x6c,
+	0x75, 0x65, 0x22, 0x37, 0x0a, 0x08, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55, 0x73, 0x12, 0x2b,
+	0x0a, 0x04, 0x70, 0x64, 0x75, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x17, 0x2e, 0x74,
+	0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x53, 0x6e,
+	0x6d, 0x70, 0x50, 0x44, 0x55, 0x52, 0x04, 0x70, 0x64, 0x75, 0x73, 0x22, 0x57, 0x0a, 0x0a, 0x53,
+	0x6e, 0x6d, 0x70, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x45, 0x72, 0x72,
+	0x6f, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x05, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x12,
+	0x33, 0x0a, 0x08, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x18, 0x02, 0x20, 0x03, 0x28,
+	0x0b, 0x32, 0x17, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e,
+	0x6d, 0x70, 0x2e, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55, 0x52, 0x08, 0x56, 0x61, 0x72, 0x69,
+	0x61, 0x62, 0x6c, 0x65, 0x2a, 0xf6, 0x02, 0x0a, 0x07, 0x41, 0x73, 0x6e, 0x31, 0x42, 0x45, 0x52,
+	0x12, 0x11, 0x0a, 0x0d, 0x45, 0x6e, 0x64, 0x4f, 0x66, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74,
+	0x73, 0x10, 0x00, 0x12, 0x0f, 0x0a, 0x0b, 0x55, 0x6e, 0x6b, 0x6e, 0x6f, 0x77, 0x6e, 0x54, 0x79,
+	0x70, 0x65, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x42, 0x6f, 0x6f, 0x6c, 0x65, 0x61, 0x6e, 0x10,
+	0x01, 0x12, 0x0b, 0x0a, 0x07, 0x49, 0x6e, 0x74, 0x65, 0x67, 0x65, 0x72, 0x10, 0x02, 0x12, 0x0d,
+	0x0a, 0x09, 0x42, 0x69, 0x74, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x10, 0x03, 0x12, 0x0f, 0x0a,
+	0x0b, 0x4f, 0x63, 0x74, 0x65, 0x74, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x10, 0x04, 0x12, 0x08,
+	0x0a, 0x04, 0x4e, 0x75, 0x6c, 0x6c, 0x10, 0x05, 0x12, 0x14, 0x0a, 0x10, 0x4f, 0x62, 0x6a, 0x65,
+	0x63, 0x74, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x66, 0x69, 0x65, 0x72, 0x10, 0x06, 0x12, 0x15,
+	0x0a, 0x11, 0x4f, 0x62, 0x6a, 0x65, 0x63, 0x74, 0x44, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74,
+	0x69, 0x6f, 0x6e, 0x10, 0x07, 0x12, 0x0d, 0x0a, 0x09, 0x49, 0x50, 0x41, 0x64, 0x64, 0x72, 0x65,
+	0x73, 0x73, 0x10, 0x40, 0x12, 0x0d, 0x0a, 0x09, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x65, 0x72, 0x33,
+	0x32, 0x10, 0x41, 0x12, 0x0b, 0x0a, 0x07, 0x47, 0x61, 0x75, 0x67, 0x65, 0x33, 0x32, 0x10, 0x42,
+	0x12, 0x0d, 0x0a, 0x09, 0x54, 0x69, 0x6d, 0x65, 0x54, 0x69, 0x63, 0x6b, 0x73, 0x10, 0x43, 0x12,
+	0x0a, 0x0a, 0x06, 0x4f, 0x70, 0x61, 0x71, 0x75, 0x65, 0x10, 0x44, 0x12, 0x0f, 0x0a, 0x0b, 0x4e,
+	0x73, 0x61, 0x70, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x10, 0x45, 0x12, 0x0d, 0x0a, 0x09,
+	0x43, 0x6f, 0x75, 0x6e, 0x74, 0x65, 0x72, 0x36, 0x34, 0x10, 0x46, 0x12, 0x0e, 0x0a, 0x0a, 0x55,
+	0x69, 0x6e, 0x74, 0x65, 0x67, 0x65, 0x72, 0x33, 0x32, 0x10, 0x47, 0x12, 0x0f, 0x0a, 0x0b, 0x4f,
+	0x70, 0x61, 0x71, 0x75, 0x65, 0x46, 0x6c, 0x6f, 0x61, 0x74, 0x10, 0x78, 0x12, 0x10, 0x0a, 0x0c,
+	0x4f, 0x70, 0x61, 0x71, 0x75, 0x65, 0x44, 0x6f, 0x75, 0x62, 0x6c, 0x65, 0x10, 0x79, 0x12, 0x11,
+	0x0a, 0x0c, 0x4e, 0x6f, 0x53, 0x75, 0x63, 0x68, 0x4f, 0x62, 0x6a, 0x65, 0x63, 0x74, 0x10, 0x80,
+	0x01, 0x12, 0x13, 0x0a, 0x0e, 0x4e, 0x6f, 0x53, 0x75, 0x63, 0x68, 0x49, 0x6e, 0x73, 0x74, 0x61,
+	0x6e, 0x63, 0x65, 0x10, 0x81, 0x01, 0x12, 0x11, 0x0a, 0x0c, 0x45, 0x6e, 0x64, 0x4f, 0x66, 0x4d,
+	0x69, 0x62, 0x56, 0x69, 0x65, 0x77, 0x10, 0x82, 0x01, 0x1a, 0x02, 0x10, 0x01, 0x32, 0xff, 0x01,
+	0x0a, 0x07, 0x43, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x12, 0x3c, 0x0a, 0x03, 0x47, 0x65, 0x74,
+	0x12, 0x17, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d,
+	0x70, 0x2e, 0x4f, 0x69, 0x64, 0x4c, 0x69, 0x73, 0x74, 0x1a, 0x1a, 0x2e, 0x74, 0x68, 0x65, 0x62,
+	0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x53, 0x6e, 0x6d, 0x70, 0x50,
+	0x61, 0x63, 0x6b, 0x65, 0x74, 0x22, 0x00, 0x12, 0x3d, 0x0a, 0x03, 0x53, 0x65, 0x74, 0x12, 0x18,
+	0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e,
+	0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55, 0x73, 0x1a, 0x1a, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69,
+	0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x61,
+	0x63, 0x6b, 0x65, 0x74, 0x22, 0x00, 0x12, 0x37, 0x0a, 0x04, 0x57, 0x61, 0x6c, 0x6b, 0x12, 0x13,
+	0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e,
+	0x4f, 0x69, 0x64, 0x1a, 0x18, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e,
+	0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55, 0x73, 0x22, 0x00, 0x12,
+	0x3e, 0x0a, 0x0a, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x57, 0x61, 0x6c, 0x6b, 0x12, 0x13, 0x2e,
+	0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73, 0x6e, 0x6d, 0x70, 0x2e, 0x4f,
+	0x69, 0x64, 0x1a, 0x17, 0x2e, 0x74, 0x68, 0x65, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x2e, 0x73,
+	0x6e, 0x6d, 0x70, 0x2e, 0x53, 0x6e, 0x6d, 0x70, 0x50, 0x44, 0x55, 0x22, 0x00, 0x30, 0x01, 0x62,
+	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+})
+
+var (
+	file_snmp_proto_rawDescOnce sync.Once
+	file_snmp_proto_rawDescData []byte
+)
+
+func file_snmp_proto_rawDescGZIP() []byte {
+	file_snmp_proto_rawDescOnce.Do(func() {
+		file_snmp_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_snmp_proto_rawDesc), len(file_snmp_proto_rawDesc)))
+	})
+	return file_snmp_proto_rawDescData
 }
 
-func (m *SnmpPDU) GetUI64() uint64 {
-	if x, ok := m.GetValue().(*SnmpPDU_UI64); ok {
-		return x.UI64
-	}
-	return 0
+var file_snmp_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_snmp_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
+var file_snmp_proto_goTypes = []any{
+	(Asn1BER)(0),       // 0: thebinary.snmp.Asn1BER
+	(*Oid)(nil),        // 1: thebinary.snmp.Oid
+	(*OidList)(nil),    // 2: thebinary.snmp.OidList
+	(*SnmpPDU)(nil),    // 3: thebinary.snmp.SnmpPDU
+	(*SnmpPDUs)(nil),   // 4: thebinary.snmp.SnmpPDUs
+	(*SnmpPacket)(nil), // 5: thebinary.snmp.SnmpPacket
+}
+var file_snmp_proto_depIdxs = []int32{
+	0, // 0: thebinary.snmp.SnmpPDU.Type:type_name -> thebinary.snmp.Asn1BER
+	3, // 1: thebinary.snmp.SnmpPDUs.pdus:type_name -> thebinary.snmp.SnmpPDU
+	3, // 2: thebinary.snmp.SnmpPacket.Variable:type_name -> thebinary.snmp.SnmpPDU
+	2, // 3: thebinary.snmp.Command.Get:input_type -> thebinary.snmp.OidList
+	4, // 4: thebinary.snmp.Command.Set:input_type -> thebinary.snmp.SnmpPDUs
+	1, // 5: thebinary.snmp.Command.Walk:input_type -> thebinary.snmp.Oid
+	1, // 6: thebinary.snmp.Command.StreamWalk:input_type -> thebinary.snmp.Oid
+	5, // 7: thebinary.snmp.Command.Get:output_type -> thebinary.snmp.SnmpPacket
+	5, // 8: thebinary.snmp.Command.Set:output_type -> thebinary.snmp.SnmpPacket
+	4, // 9: thebinary.snmp.Command.Walk:output_type -> thebinary.snmp.SnmpPDUs
+	3, // 10: thebinary.snmp.Command.StreamWalk:output_type -> thebinary.snmp.SnmpPDU
+	7, // [7:11] is the sub-list for method output_type
+	3, // [3:7] is the sub-list for method input_type
+	3, // [3:3] is the sub-list for extension type_name
+	3, // [3:3] is the sub-list for extension extendee
+	0, // [0:3] is the sub-list for field type_name
 }
 
-func (m *SnmpPDU) GetStr() string {
-	if x, ok := m.GetValue().(*SnmpPDU_Str); ok {
-		return x.Str
+func init() { file_snmp_proto_init() }
+func file_snmp_proto_init() {
+	if File_snmp_proto != nil {
+		return
 	}
-	return ""
-}
-
-// XXX_OneofWrappers is for the internal use of the proto package.
-func (*SnmpPDU) XXX_OneofWrappers() []interface{} {
-	return []interface{}{
+	file_snmp_proto_msgTypes[2].OneofWrappers = []any{
 		(*SnmpPDU_I32)(nil),
 		(*SnmpPDU_I64)(nil),
 		(*SnmpPDU_UI32)(nil),
 		(*SnmpPDU_UI64)(nil),
 		(*SnmpPDU_Str)(nil),
 	}
-}
-
-//
-// Represents multiple SNMP PDU
-type SnmpPDUs struct {
-	Pdus                 []*SnmpPDU `protobuf:"bytes,1,rep,name=pdus,proto3" json:"pdus,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}   `json:"-"`
-	XXX_unrecognized     []byte     `json:"-"`
-	XXX_sizecache        int32      `json:"-"`
-}
-
-func (m *SnmpPDUs) Reset()         { *m = SnmpPDUs{} }
-func (m *SnmpPDUs) String() string { return proto.CompactTextString(m) }
-func (*SnmpPDUs) ProtoMessage()    {}
-func (*SnmpPDUs) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{3}
-}
-
-func (m *SnmpPDUs) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_SnmpPDUs.Unmarshal(m, b)
-}
-func (m *SnmpPDUs) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_SnmpPDUs.Marshal(b, m, deterministic)
-}
-func (m *SnmpPDUs) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_SnmpPDUs.Merge(m, src)
-}
-func (m *SnmpPDUs) XXX_Size() int {
-	return xxx_messageInfo_SnmpPDUs.Size(m)
-}
-func (m *SnmpPDUs) XXX_DiscardUnknown() {
-	xxx_messageInfo_SnmpPDUs.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_SnmpPDUs proto.InternalMessageInfo
-
-func (m *SnmpPDUs) GetPdus() []*SnmpPDU {
-	if m != nil {
-		return m.Pdus
-	}
-	return nil
-}
-
-type SnmpPacket struct {
-	Error                uint32     `protobuf:"varint,1,opt,name=Error,proto3" json:"Error,omitempty"`
-	Variable             []*SnmpPDU `protobuf:"bytes,2,rep,name=Variable,proto3" json:"Variable,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}   `json:"-"`
-	XXX_unrecognized     []byte     `json:"-"`
-	XXX_sizecache        int32      `json:"-"`
-}
-
-func (m *SnmpPacket) Reset()         { *m = SnmpPacket{} }
-func (m *SnmpPacket) String() string { return proto.CompactTextString(m) }
-func (*SnmpPacket) ProtoMessage()    {}
-func (*SnmpPacket) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7dcafed58c309cd0, []int{4}
-}
-
-func (m *SnmpPacket) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_SnmpPacket.Unmarshal(m, b)
-}
-func (m *SnmpPacket) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_SnmpPacket.Marshal(b, m, deterministic)
-}
-func (m *SnmpPacket) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_SnmpPacket.Merge(m, src)
-}
-func (m *SnmpPacket) XXX_Size() int {
-	return xxx_messageInfo_SnmpPacket.Size(m)
-}
-func (m *SnmpPacket) XXX_DiscardUnknown() {
-	xxx_messageInfo_SnmpPacket.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_SnmpPacket proto.InternalMessageInfo
-
-func (m *SnmpPacket) GetError() uint32 {
-	if m != nil {
-		return m.Error
-	}
-	return 0
-}
-
-func (m *SnmpPacket) GetVariable() []*SnmpPDU {
-	if m != nil {
-		return m.Variable
-	}
-	return nil
-}
-
-func init() {
-	proto.RegisterEnum("thebinary.snmp.Asn1BER", Asn1BER_name, Asn1BER_value)
-	proto.RegisterType((*Oid)(nil), "thebinary.snmp.Oid")
-	proto.RegisterType((*OidList)(nil), "thebinary.snmp.OidList")
-	proto.RegisterType((*SnmpPDU)(nil), "thebinary.snmp.SnmpPDU")
-	proto.RegisterType((*SnmpPDUs)(nil), "thebinary.snmp.SnmpPDUs")
-	proto.RegisterType((*SnmpPacket)(nil), "thebinary.snmp.SnmpPacket")
-}
-
-func init() { proto.RegisterFile("snmp.proto", fileDescriptor_7dcafed58c309cd0) }
-
-var fileDescriptor_7dcafed58c309cd0 = []byte{
-	// 623 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x84, 0x54, 0xdf, 0x6f, 0xda, 0x3a,
-	0x14, 0xc6, 0x24, 0x90, 0xf6, 0xb4, 0x70, 0x5d, 0xb7, 0x57, 0x8d, 0x90, 0xae, 0x84, 0xf2, 0x84,
-	0xee, 0x24, 0xb4, 0x01, 0xa2, 0x2f, 0xdb, 0xb4, 0xf2, 0xa3, 0x2d, 0xd2, 0x06, 0x55, 0x28, 0xed,
-	0xb3, 0x21, 0x6e, 0xeb, 0x11, 0xec, 0xcc, 0x71, 0xd4, 0xf5, 0x6d, 0xdb, 0xbf, 0xb5, 0xff, 0x6b,
-	0xaf, 0x9b, 0xec, 0xd0, 0x69, 0x53, 0xcb, 0xf6, 0x76, 0xce, 0xe7, 0xef, 0x3b, 0xfe, 0xce, 0x39,
-	0x4e, 0x00, 0x52, 0xb1, 0x4a, 0x9a, 0x89, 0x92, 0x5a, 0x92, 0xaa, 0xbe, 0x65, 0x73, 0x2e, 0xa8,
-	0xba, 0x6f, 0x1a, 0x34, 0x38, 0x04, 0x67, 0xc2, 0x23, 0x82, 0xc1, 0x91, 0x3c, 0xf2, 0x51, 0x1d,
-	0x35, 0xb6, 0x43, 0x13, 0x06, 0xff, 0x81, 0x37, 0xe1, 0xd1, 0x5b, 0x9e, 0x6a, 0x42, 0xc0, 0x95,
-	0x3c, 0x4a, 0x7d, 0x54, 0x77, 0x1a, 0xdb, 0xa1, 0x8d, 0x83, 0xaf, 0x08, 0xbc, 0xa9, 0x58, 0x25,
-	0xe7, 0x83, 0x99, 0x39, 0x1f, 0xd3, 0x15, 0x5b, 0xab, 0x6d, 0x4c, 0x9e, 0x81, 0x7b, 0x71, 0x9f,
-	0x30, 0xbf, 0x58, 0x47, 0x8d, 0x6a, 0xeb, 0xb0, 0xf9, 0xfb, 0xb5, 0xcd, 0xe3, 0x54, 0xbc, 0xe8,
-	0x0d, 0xc3, 0xd0, 0x92, 0x08, 0x01, 0x67, 0xd4, 0x6e, 0xf9, 0x4e, 0x1d, 0x35, 0x4a, 0x67, 0x85,
-	0xd0, 0x24, 0x16, 0xeb, 0x76, 0x7c, 0xb7, 0x8e, 0x1a, 0x8e, 0xc5, 0xba, 0x1d, 0x72, 0x00, 0xee,
-	0xcc, 0x10, 0x4b, 0x75, 0xd4, 0xa8, 0x9c, 0x15, 0x42, 0x9b, 0xe5, 0x68, 0xb7, 0xe3, 0x97, 0xeb,
-	0xa8, 0xe1, 0xe6, 0x68, 0xb7, 0x63, 0xf4, 0x53, 0xad, 0x7c, 0xcf, 0x78, 0x32, 0xfa, 0xa9, 0x56,
-	0x3d, 0x0f, 0x4a, 0x97, 0x34, 0xce, 0x58, 0x70, 0x04, 0x5b, 0x6b, 0xf3, 0xa9, 0x71, 0x9a, 0x44,
-	0x59, 0xde, 0xdd, 0xce, 0x63, 0xa7, 0x6b, 0x5e, 0x68, 0x49, 0xc1, 0x15, 0x80, 0x05, 0xe8, 0x62,
-	0xc9, 0x34, 0x39, 0x80, 0xd2, 0x50, 0x29, 0xa9, 0x6c, 0xe7, 0x95, 0x30, 0x4f, 0x48, 0x1b, 0xb6,
-	0x2e, 0xa9, 0xe2, 0x74, 0x1e, 0x9b, 0xf6, 0xff, 0x58, 0xf4, 0x27, 0xf1, 0xff, 0x6f, 0x45, 0xf0,
-	0xd6, 0x43, 0x21, 0x7b, 0x50, 0x19, 0x8a, 0x68, 0x72, 0xdd, 0x97, 0x42, 0x33, 0xa1, 0x53, 0x5c,
-	0x20, 0xff, 0xc0, 0xce, 0x4c, 0x2c, 0x85, 0xbc, 0x13, 0x66, 0x60, 0xb8, 0x40, 0x76, 0xc0, 0xeb,
-	0x49, 0x19, 0x33, 0x2a, 0x30, 0x32, 0xc9, 0x48, 0x68, 0x76, 0xc3, 0x14, 0x2e, 0x92, 0x0a, 0x6c,
-	0xf7, 0xb8, 0x9e, 0x6a, 0xc5, 0xc5, 0x0d, 0x76, 0x8c, 0x72, 0xb2, 0xd0, 0xec, 0x01, 0x70, 0xc9,
-	0x16, 0xb8, 0xe3, 0x2c, 0x8e, 0x71, 0x89, 0x1c, 0x00, 0x9e, 0xcc, 0xdf, 0xb3, 0x85, 0x1e, 0x45,
-	0x4c, 0x68, 0x7e, 0xcd, 0x99, 0xc2, 0x65, 0xf2, 0x2f, 0xec, 0xe5, 0xe8, 0x80, 0xa5, 0x0b, 0xc5,
-	0x13, 0xcd, 0xa5, 0xc0, 0x9e, 0x29, 0x3b, 0x3a, 0x3f, 0x8e, 0x22, 0xc5, 0xd2, 0x14, 0xbf, 0x31,
-	0x69, 0x5f, 0x66, 0x42, 0x33, 0xd5, 0x6e, 0xe1, 0x63, 0xe3, 0xe0, 0x94, 0x66, 0x37, 0xac, 0xdd,
-	0xc2, 0x3d, 0x73, 0x76, 0xc1, 0x57, 0xec, 0x82, 0x2f, 0x96, 0x29, 0xee, 0x13, 0x80, 0xf2, 0x24,
-	0xa1, 0x1f, 0x32, 0x86, 0x07, 0xc6, 0xcd, 0x38, 0xa5, 0xc9, 0x43, 0x9d, 0xe1, 0x2f, 0x75, 0xba,
-	0x1d, 0x7c, 0x42, 0xaa, 0x00, 0x33, 0x9e, 0xb7, 0xd2, 0x6e, 0xe1, 0x53, 0xeb, 0xde, 0x6a, 0x4f,
-	0x62, 0x49, 0x35, 0xfe, 0x48, 0x30, 0xec, 0xe6, 0xc0, 0x40, 0x66, 0xf3, 0x98, 0xe1, 0x7b, 0xb2,
-	0x07, 0xbb, 0x63, 0x39, 0xcd, 0x16, 0xb7, 0xb9, 0x6b, 0xfc, 0x09, 0x91, 0x7d, 0xa8, 0xe6, 0xd0,
-	0x48, 0xa4, 0x9a, 0x8a, 0x05, 0xc3, 0x9f, 0x91, 0xe1, 0xd9, 0xa9, 0xbe, 0xe3, 0xf3, 0x4b, 0xce,
-	0xee, 0xf0, 0x17, 0x54, 0x2b, 0x62, 0xd4, 0xfa, 0x8e, 0xc0, 0xeb, 0xcb, 0xd5, 0x8a, 0x8a, 0x88,
-	0xbc, 0x04, 0xe7, 0x94, 0x69, 0xf2, 0x68, 0x5d, 0xeb, 0x0f, 0xa1, 0x56, 0x7b, 0x72, 0x8f, 0xf6,
-	0x2d, 0x04, 0x05, 0xf2, 0x0a, 0x9c, 0x29, 0xd3, 0xc4, 0xdf, 0xb0, 0xec, 0xf4, 0x2f, 0xf2, 0x23,
-	0x70, 0xaf, 0x68, 0xbc, 0x24, 0xfb, 0x4f, 0xdc, 0x5e, 0xdb, 0x58, 0x34, 0x28, 0x90, 0xd7, 0x00,
-	0x53, 0xad, 0x18, 0x5d, 0x6d, 0x96, 0x6f, 0x7a, 0x80, 0x41, 0xe1, 0x39, 0x9a, 0x97, 0xed, 0x9f,
-	0xa1, 0xfd, 0x23, 0x00, 0x00, 0xff, 0xff, 0x9c, 0xcf, 0x22, 0xa1, 0x27, 0x04, 0x00, 0x00,
-}
-
-// Reference imports to suppress errors if they are not otherwise used.
-var _ context.Context
-var _ grpc.ClientConn
-
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion4
-
-// CommandClient is the client API for Command service.
-//
-// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
-type CommandClient interface {
-	//SNMP Get given the list of OIDs
-	Get(ctx context.Context, in *OidList, opts ...grpc.CallOption) (*SnmpPacket, error)
-	//SNMP Set
-	Set(ctx context.Context, in *SnmpPDUs, opts ...grpc.CallOption) (*SnmpPacket, error)
-	//SNMP Walk and return all variables in one shot
-	Walk(ctx context.Context, in *Oid, opts ...grpc.CallOption) (*SnmpPDUs, error)
-	//Stream each SNMP PDU while running SNMP WALK from the given OID.
-	StreamWalk(ctx context.Context, in *Oid, opts ...grpc.CallOption) (Command_StreamWalkClient, error)
-}
-
-type commandClient struct {
-	cc *grpc.ClientConn
-}
-
-func NewCommandClient(cc *grpc.ClientConn) CommandClient {
-	return &commandClient{cc}
-}
-
-func (c *commandClient) Get(ctx context.Context, in *OidList, opts ...grpc.CallOption) (*SnmpPacket, error) {
-	out := new(SnmpPacket)
-	err := c.cc.Invoke(ctx, "/thebinary.snmp.Command/Get", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *commandClient) Set(ctx context.Context, in *SnmpPDUs, opts ...grpc.CallOption) (*SnmpPacket, error) {
-	out := new(SnmpPacket)
-	err := c.cc.Invoke(ctx, "/thebinary.snmp.Command/Set", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *commandClient) Walk(ctx context.Context, in *Oid, opts ...grpc.CallOption) (*SnmpPDUs, error) {
-	out := new(SnmpPDUs)
-	err := c.cc.Invoke(ctx, "/thebinary.snmp.Command/Walk", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *commandClient) StreamWalk(ctx context.Context, in *Oid, opts ...grpc.CallOption) (Command_StreamWalkClient, error) {
-	stream, err := c.cc.NewStream(ctx, &_Command_serviceDesc.Streams[0], "/thebinary.snmp.Command/StreamWalk", opts...)
-	if err != nil {
-		return nil, err
-	}
-	x := &commandStreamWalkClient{stream}
-	if err := x.ClientStream.SendMsg(in); err != nil {
-		return nil, err
-	}
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	return x, nil
-}
-
-type Command_StreamWalkClient interface {
-	Recv() (*SnmpPDU, error)
-	grpc.ClientStream
-}
-
-type commandStreamWalkClient struct {
-	grpc.ClientStream
-}
-
-func (x *commandStreamWalkClient) Recv() (*SnmpPDU, error) {
-	m := new(SnmpPDU)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-// CommandServer is the server API for Command service.
-type CommandServer interface {
-	//SNMP Get given the list of OIDs
-	Get(context.Context, *OidList) (*SnmpPacket, error)
-	//SNMP Set
-	Set(context.Context, *SnmpPDUs) (*SnmpPacket, error)
-	//SNMP Walk and return all variables in one shot
-	Walk(context.Context, *Oid) (*SnmpPDUs, error)
-	//Stream each SNMP PDU while running SNMP WALK from the given OID.
-	StreamWalk(*Oid, Command_StreamWalkServer) error
-}
-
-// UnimplementedCommandServer can be embedded to have forward compatible implementations.
-type UnimplementedCommandServer struct {
-}
-
-func (*UnimplementedCommandServer) Get(ctx context.Context, req *OidList) (*SnmpPacket, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Get not implemented")
-}
-func (*UnimplementedCommandServer) Set(ctx context.Context, req *SnmpPDUs) (*SnmpPacket, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Set not implemented")
-}
-func (*UnimplementedCommandServer) Walk(ctx context.Context, req *Oid) (*SnmpPDUs, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Walk not implemented")
-}
-func (*UnimplementedCommandServer) StreamWalk(req *Oid, srv Command_StreamWalkServer) error {
-	return status.Errorf(codes.Unimplemented, "method StreamWalk not implemented")
-}
-
-func RegisterCommandServer(s *grpc.Server, srv CommandServer) {
-	s.RegisterService(&_Command_serviceDesc, srv)
-}
-
-func _Command_Get_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(OidList)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(CommandServer).Get(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/thebinary.snmp.Command/Get",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(CommandServer).Get(ctx, req.(*OidList))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Command_Set_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(SnmpPDUs)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(CommandServer).Set(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/thebinary.snmp.Command/Set",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(CommandServer).Set(ctx, req.(*SnmpPDUs))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Command_Walk_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(Oid)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(CommandServer).Walk(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/thebinary.snmp.Command/Walk",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(CommandServer).Walk(ctx, req.(*Oid))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Command_StreamWalk_Handler(srv interface{}, stream grpc.ServerStream) error {
-	m := new(Oid)
-	if err := stream.RecvMsg(m); err != nil {
-		return err
-	}
-	return srv.(CommandServer).StreamWalk(m, &commandStreamWalkServer{stream})
-}
-
-type Command_StreamWalkServer interface {
-	Send(*SnmpPDU) error
-	grpc.ServerStream
-}
-
-type commandStreamWalkServer struct {
-	grpc.ServerStream
-}
-
-func (x *commandStreamWalkServer) Send(m *SnmpPDU) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-var _Command_serviceDesc = grpc.ServiceDesc{
-	ServiceName: "thebinary.snmp.Command",
-	HandlerType: (*CommandServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "Get",
-			Handler:    _Command_Get_Handler,
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_snmp_proto_rawDesc), len(file_snmp_proto_rawDesc)),
+			NumEnums:      1,
+			NumMessages:   5,
+			NumExtensions: 0,
+			NumServices:   1,
 		},
-		{
-			MethodName: "Set",
-			Handler:    _Command_Set_Handler,
-		},
-		{
-			MethodName: "Walk",
-			Handler:    _Command_Walk_Handler,
-		},
-	},
-	Streams: []grpc.StreamDesc{
-		{
-			StreamName:    "StreamWalk",
-			Handler:       _Command_StreamWalk_Handler,
-			ServerStreams: true,
-		},
-	},
-	Metadata: "snmp.proto",
+		GoTypes:           file_snmp_proto_goTypes,
+		DependencyIndexes: file_snmp_proto_depIdxs,
+		EnumInfos:         file_snmp_proto_enumTypes,
+		MessageInfos:      file_snmp_proto_msgTypes,
+	}.Build()
+	File_snmp_proto = out.File
+	file_snmp_proto_goTypes = nil
+	file_snmp_proto_depIdxs = nil
 }
