@@ -1,8 +1,10 @@
 package server
 
-//go:generate protoc -I ../protobuf --go_out=plugins=grpc:../protobuf ../protobuf/snmp.proto
 //go:generate protoc --doc_out=../protobuf --doc_opt=markdown,../protobuf/README.md ../protobuf/snmp.proto
-import "github.com/soniah/gosnmp"
+import (
+	"github.com/soniah/gosnmp"
+	"github.com/thebinary/go-snmp-grpc/protobuf"
+)
 
 // Logger should implement logging functionality
 type Logger interface{}
@@ -14,4 +16,6 @@ type CommandServer struct {
 	DefaultCommunity string
 	Logger           Logger
 	SNMPLogger       gosnmp.Logger
+
+	protobuf.UnimplementedCommandServer
 }
